refactor field code locations
