+++ dynamic_rest/fields/__init__.py
@@ -1,2 +1,4 @@
-from dynamic_rest.fields.fields import *  # noqa
-from dynamic_rest.fields.generic import *  # noqa
+# flake8: noqa
+from .base import DynamicField, DynamicMethodField, CountField
+from .relation import DynamicRelationField
+from .generic import DynamicGenericRelationField
