+++ dynamic_rest/viewsets.py
@@ -1,5 +1,6 @@
 """This module contains custom viewset classes."""
 import csv
+import inflection
 from django.core.exceptions import ObjectDoesNotExist
 from django.http import QueryDict
 from django.utils import six
@@ -19,6 +20,24 @@ from dynamic_rest.utils import is_truthy
 
 UPDATE_REQUEST_METHODS = ('PUT', 'PATCH', 'POST')
 DELETE_REQUEST_METHOD = 'DELETE'
+
+
+def get_view_name(view_cls, suffix=None):
+    serializer_class = getattr(view_cls, 'serializer_class', None)
+    suffix = suffix or ''
+    if serializer_class:
+        serializer = view_cls.serializer_class()
+        if suffix.lower() == 'list':
+            name = serializer.get_plural_name()
+        else:
+            name = serializer.get_name()
+    else:
+        name = view_cls.__name__
+        name = (
+            inflection.pluralize(name)
+            if suffix.lower() == 'list' else name
+        )
+    return name.title()
 
 
 class QueryParams(QueryDict):
